pyproject.toml
[tool.poetry]
name = "lelikov_diff"
version = "0.2.2"
description = ""
authors = ["Alexandr Lelikov <lelikovas@gmail.com>"]
packages = [
  { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.8"
PyYAML = "^5.2"

[tool.poetry.dev-dependencies]
flake8 = "^3.7.9"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
